apps/api/requirements.txt
fastapi==0.116.2
uvicorn==0.35.0
requests==2.32.5
livekit==1.0.13
python-dotenv==1.0.0